Fix incorrect file naming for tests
